Refactor lottomax_route, lottomax_repository, and lottomax_factory
